# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "SwiftCloud API",
    "description": "A fabulous API to settle your cravings for Taylor Swift music statistics!"
  },
  "host": "swiftcloud.ckret.net",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/api/artists": {
      "get": {
        "tags": [
          "Artists"
        ],
        "summary": "Gets all artists",
        "description": "Gets all artists",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/artist/{id}": {
      "get": {
        "tags": [
          "Artists"
        ],
        "summary": "Gets artist by id",
        "description": "Gets an artists with the specified id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The id to find"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/artist/name/{name}": {
      "get": {
        "tags": [
          "Artists"
        ],
        "summary": "Gets artist by name",
        "description": "Gets an artists with the specified name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The name to find"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/writers": {
      "get": {
        "tags": [
          "Writers"
        ],
        "summary": "Gets all writers",
        "description": "Gets all writers",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/writer/{id}": {
      "get": {
        "tags": [
          "Writers"
        ],
        "summary": "Gets writer by id",
        "description": "Gets an writers with the specified id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The id to find"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/writer/name/{name}": {
      "get": {
        "tags": [
          "Writers"
        ],
        "summary": "Gets writer by name",
        "description": "Gets an writers with the specified name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The name to find"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/songs": {
      "get": {
        "tags": [
          "Songs"
        ],
        "summary": "Gets all songs",
        "description": "Gets all songs with or without relations",
        "parameters": [
          {
            "name": "rel",
            "in": "query",
            "type": "boolean",
            "description": "Get relational data"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/song/{id}": {
      "get": {
        "tags": [
          "Songs"
        ],
        "summary": "Gets song by id",
        "description": "Finds a song with the specified id, with or without relations",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "rel",
            "in": "query",
            "type": "boolean",
            "description": "Get relational data"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/song/name/{name}": {
      "get": {
        "tags": [
          "Songs"
        ],
        "summary": "Gets song by name",
        "description": "Finds a song with the specified name, with or without relations",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "rel",
            "in": "query",
            "type": "boolean",
            "description": "Get relational data"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/songs/year/{year}": {
      "get": {
        "tags": [
          "Songs"
        ],
        "summary": "All songs for the specified year",
        "description": "Gets all songs for the specified year, with or without relations",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "rel",
            "in": "query",
            "type": "boolean",
            "description": "Get relational data"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/songs/popular/{year}": {
      "get": {
        "tags": [
          "Songs"
        ],
        "summary": "Most listened songs for a given year",
        "description": "Gets the specified number of songs ordered by number of times listened to during the specified year, with or without relations",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "number",
            "description": "Limit the result to # entities"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/songs/popular/{year}/{month}": {
      "get": {
        "tags": [
          "Songs"
        ],
        "summary": "Most listened songs for a given year and month",
        "description": "Gets the specified number of songs ordered by number of times listened to during the specified year and month, with or without relations",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "number",
            "description": "Limit the result to # entities"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}